Enable full C++ ranges for collections using fast_iterator
