+++ cppwinrt/cppwinrt.vcxproj.filters
@@ -163,6 +163,9 @@
     <ClInclude Include="..\strings\base_coroutine_system_winui.h">
       <Filter>strings</Filter>
     </ClInclude>
+    <ClInclude Include="..\strings\base_iterator.h">
+      <Filter>strings</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="$(OutDir)version.rc" />
